fix: fix version check
