+++ fe-app-podkop/src/helpers/removeVersionPrefix.ts
@@ -1,0 +1,1 @@
+export const removeVersionPrefix = (v: string) => v.replace(/^v/, '');
